Use --no-ignore-vcs in default searches
Newer versions of virtualenv add a .gitignore file containing a line
with * in it, which ignores all files under the virtualenv. If the
virtualenv happens to be under a git repository (such as when a user's
home directory is tracked by git), fd will respect the gitignore files
it sees in virtualenvs, causing venv-selector's default searches to fail
to work as intended.