Fix overwriting client settings on notify for pylsp
When pylsp gets settings via didChangeConfiguration it simply replaces
the former settings by the given settings. Therefore instead of just
giving the changed settings, we provide all the settings which we
updated in a previous steps. This prevents the settings from being
cleared.

This should be save to do, since pylsp is the only given lsp server
which uses the 'use_settings_directly' variable and therefore should be
the only lsp affected by this change.