Add back types for user config
This adds back the types that were originally present for the config.
Instead of duplicating the details in a class at the top, the fields are
read automatically from the declaration of the default settings.

There is no way to mark fields as deprecated, so deprecated fields have
deliberately not been listed. This means that luaLS will have warnings
that it doesn't know about the fields inside the `merge_user_settings`
function, but it also means that those options won't be suggested in the
autocomplete suggestions.